backend/pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[project]
name = "bi-bot"
version = "0.1.0"
description = "AI-powered business intelligence bot for natural language database queries"
requires-python = ">=3.11"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
google-generativeai = "^0.3.0"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.23"
pymongo = "^4.6.0"
pyjwt = "^2.8.1"
sqlparse = "^0.4.4"
slowapi = "^0.1.9"
httpx = "^0.25.1"
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"

[tool.poetry.group.dev.dependencies]
black = "^23.12.0"
ruff = "^0.1.8"
mypy = "^1.7.1"

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
select = ["E", "F", "W"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
